Added menu_handler.rb file
